chore(ui): refacto & simplification
wip


wip


wip


wip
