+++ ui/components/Filters/DateWithTooltipSelector.jsx
@@ -2,7 +2,7 @@ import { HStack, Text, Tooltip } from "@chakra-ui/react";
 import React from "react";
 import { MonthSelect } from "../index.jsx";
 import { InfoLine } from "../../theme/components/icons/InfoLine.js";
-import { useSimpleFiltersContext } from "../../modules/mon-espace/landing/LandingOrganisme/SimpleFiltersContext.js";
+import { useSimpleFiltersContext } from "../../modules/mon-espace/landing/common/SimpleFiltersContext.js";
 
 const DateWithTooltipSelector = ({ ...props }) => {
   const { filtersValues, filtersSetters } = useSimpleFiltersContext();
