+++ ui/modules/mon-espace/landing/LandingOrganisme/LandingOrganisme.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Heading } from "@chakra-ui/react";
+import { Heading, Stack } from "@chakra-ui/react";
 import { useEspace } from "../../../../hooks/useEspace";
 import OrganismeInfo from "./components/OrganismeInfo";
 
@@ -7,14 +7,14 @@ const LandingOrganisme = () => {
   const { isMonOrganismePages, isOrganismePages } = useEspace();
 
   return (
-    <>
-      <Heading textStyle="h2" color="grey.800" mt={5}>
+    <Stack spacing="2w">
+      <Heading textStyle="h2" color="grey.800">
         {isMonOrganismePages && `Bienvenue sur votre tableau de bord`}
         {isOrganismePages && `Bienvenue sur le tableau de bord de :`}
       </Heading>
 
       <OrganismeInfo />
-    </>
+    </Stack>
   );
 };
