+++ ui/modules/mon-espace/landing/LandingReseau.jsx
@@ -1,13 +1,19 @@
 import React from "react";
-import { Heading } from "@chakra-ui/react";
+import { Heading, Stack } from "@chakra-ui/react";
+import { SimpleFiltersProvider } from "./common/SimpleFiltersContext.js";
+import IndicateursInfo from "./common/IndicateursInfos.jsx";
 
 const LandingReseau = () => {
   return (
-    <>
-      <Heading textStyle="h2" color="grey.800" mt={5}>
+    <Stack spacing="2w">
+      <Heading textStyle="h2" color="grey.800">
         Bienvenue sur votre tableau de bord [RESEAU]
       </Heading>
-    </>
+
+      <SimpleFiltersProvider>
+        <IndicateursInfo />
+      </SimpleFiltersProvider>
+    </Stack>
   );
 };
